% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcorr.R
\name{get_pcor}
\alias{get_pcor}
\title{Partial correlation test with covariates and x y variables}
\usage{
get_pcor(anth = anthro, clean = clean_dat, data = dat, covs)
}
\arguments{
\item{anth}{string vector variable}

\item{clean}{a dataframe variable}

\item{data}{a dataframe variable}

\item{covs}{a string vector variable for all covariates}
}
\value{
list of two dataframes (partial correlation table and a p-value count table)
}
\description{
Partial correlation test with covariates and x y variables
}
