% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcorr.R
\name{table_display}
\alias{table_display}
\title{Display a table in rmarkdown file only}
\usage{
table_display(df, title, row_name = TRUE)
}
\arguments{
\item{df}{a dataframe (table) variable}

\item{title}{a string variable}

\item{row_name}{a boolean variable}
}
\value{
nothing returns, just to display a table in rmarkdown
}
\description{
Display a table in rmarkdown file only
}
